requirements.txt
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pillow==12.0.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
sympy==1.14.0
torch==2.11.0.dev20260123+cu128
torchaudio==2.11.0.dev20260123+cu128
torchvision==0.25.0.dev20260123+cu128
typing_extensions==4.15.0
